Improve create API from OpenAPI URL
