+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/site/public/locales/en.json
@@ -4,12 +4,17 @@
   "Apis.Create.Components.APIInputForm.api.name.placeholder": "eTicketing",
   "Apis.Create.Components.APIInputForm.apiproduct.name.placeholder": "eTicketing",
   "Apis.Create.Components.APIInputForm.context.placeholder": "/events/{id}/sales",
+  "Apis.Create.Default.APICreateDefault.cancel": "Cancel",
   "Apis.Create.Default.APICreateDefault.gateway.url": "Gateway_URL/",
+  "Apis.Create.Default.APICreateDefault.heading": "Create an API",
+  "Apis.Create.Default.APICreateDefault.sub.heading": "Create an API providing Name, Version and Context parameters and optionally backend endpoint and bushiness plans",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.cancel": "Cancel",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.create.api": "Create API",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.create.failed": "API creation failed. Code: {errCode}, Reason(s): \"{errMessage}\"",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.create.success": "{apiAndVersion} created successfully.",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.error.empty": "OpenAPI URL",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.finish": "Finish",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.heading": "Create an API using OpenAPI definition",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.import.api": "Import an OpenAPI Definition",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.import.api.help.file": "Provide an api definition file",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.import.api.help.url": "Provide a url for the api deninition",
@@ -17,6 +22,17 @@
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.openapi.invalid.input.method": "Invalid Input Method. Either URL or File should be present.",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.openapi.validation.failed": "OpenAPI definition validation failed. Reason(s): [{failedReasons}]",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.select.openapi": "Select OpenAPI Definition",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.sub.heading": "Use an existing OpenAPI definition file or URL to create an API in WSO2 API Manager.",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.wizard.one": "Provide OpenAPI",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.wizard.two": "Create API",
+  "Apis.Create.WSDL.ApiCreateWSDL.cancel": "Cancel",
+  "Apis.Create.WSDL.ApiCreateWSDL.heading": "Create an API using WSDL",
+  "Apis.Create.WSDL.ApiCreateWSDL.sub.heading": "Use an existing SOAP endpoint to create a managed API. Import the WSDL of the SOAP service.",
+  "Apis.Create.WSDL.Steps.DefaultAPIForm.context": "Context",
+  "Apis.Create.WSDL.Steps.DefaultAPIForm.name": "Name",
+  "Apis.Create.WSDL.Steps.DefaultAPIForm.version": "Version",
+  "Apis.Create.WSDL.Steps.ProvideWSDL.Input.type": "Input type",
+  "Apis.Create.WSDL.Steps.ProvideWSDL.implementation.type": "Implementation type",
   "Apis.Create.WSDL.Steps.ProvideWSDL.uploaded.file": "File",
   "Apis.Create.WSDL.Steps.ProvideWSDL.uploaded.url": "URL",
   "Apis.Details.APIDefinition.APIDefinition.api.definition": "API Definition",
@@ -248,6 +264,9 @@
   "Apis.Details.Endpoints.Prototype.PrototypeEndpoints.endpoint": "Endpoint",
   "Apis.Details.Endpoints.Prototype.PrototypeEndpoints.endpoint.implementation.method": "Endpoint Implementation Method",
   "Apis.Details.Endpoints.Prototype.PrototypeEndpoints.inline": "Inline",
+  "Apis.Details.EndpointsNew.Endpoints.cancel": "Cancel",
+  "Apis.Details.EndpointsNew.Endpoints.endpoints.header": "Endpoints",
+  "Apis.Details.EndpointsNew.Endpoints.save": "Save",
   "Apis.Details.LifeCycle.LifeCycle.history": "History",
   "Apis.Details.LifeCycle.LifeCycle.lifecycle": "Lifecycle",
   "Apis.Details.LifeCycle.LifeCycleHistory.action": "Action",
@@ -288,6 +307,9 @@
   "Apis.Details.NewOverview.Configuration.authorization.header.tooltip": "A custom authorization header can be defined as a replacement to the default Authorization header used to send a request. If a value is specified here, it will be used as the header field to send the access tokenin a request to consume the API",
   "Apis.Details.NewOverview.Configuration.configuration": "Configuration",
   "Apis.Details.NewOverview.Configuration.context": "Context",
+  "Apis.Details.NewOverview.Configuration.cors.configuration": "CORS Configuration",
+  "Apis.Details.NewOverview.Configuration.cors.disabled": "DISABLED",
+  "Apis.Details.NewOverview.Configuration.cors.enabled": "ENABLED",
   "Apis.Details.NewOverview.Configuration.created.time": "Created Time",
   "Apis.Details.NewOverview.Configuration.default.version": "Default Version",
   "Apis.Details.NewOverview.Configuration.description": "Description",
@@ -415,10 +437,6 @@
   "Apis.Details.Resources.Resources.assign.advanced.throttling.Resource": "Apply per Resource",
   "Apis.Details.Resources.Resources.assign.advanced.throttling.perApi": "Apply per API",
   "Apis.Details.Resources.Resources.assign.advanced.throttling.policies": "Assign advanced throttling policies",
-  "Apis.Details.Resources.Resources.assign.global.scope.for.api": "Assign Global Scope for API",
-  "Apis.Details.Resources.Resources.assign.global.scopes.for.api.button": "Assign global scopes for API",
-  "Apis.Details.Resources.Resources.assign.global.scopes.for.api.input": "Assign Global Scopes for API",
-  "Apis.Details.Resources.Resources.assign.global.scopes.for.api.title": "Assign Global Scopes for API",
   "Apis.Details.Resources.Resources.assign.policies": "Assign Advanced Throttling Policies",
   "Apis.Details.Resources.Resources.cancel": "Cancel",
   "Apis.Details.Resources.Resources.delete.selected": "Delete Selected",
