+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/site/public/locales/fr.json
@@ -4,12 +4,17 @@
   "Apis.Create.Components.APIInputForm.api.name.placeholder": "",
   "Apis.Create.Components.APIInputForm.apiproduct.name.placeholder": "",
   "Apis.Create.Components.APIInputForm.context.placeholder": "",
+  "Apis.Create.Default.APICreateDefault.cancel": "",
   "Apis.Create.Default.APICreateDefault.gateway.url": "",
+  "Apis.Create.Default.APICreateDefault.heading": "",
+  "Apis.Create.Default.APICreateDefault.sub.heading": "",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.cancel": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.create.api": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.create.failed": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.create.success": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.error.empty": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.finish": "",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.heading": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.import.api": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.import.api.help.file": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.import.api.help.url": "",
@@ -17,6 +22,17 @@
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.openapi.invalid.input.method": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.openapi.validation.failed": "",
   "Apis.Create.OpenAPI.ApiCreateOpenAPI.select.openapi": "",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.sub.heading": "",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.wizard.one": "",
+  "Apis.Create.OpenAPI.ApiCreateOpenAPI.wizard.two": "",
+  "Apis.Create.WSDL.ApiCreateWSDL.cancel": "",
+  "Apis.Create.WSDL.ApiCreateWSDL.heading": "",
+  "Apis.Create.WSDL.ApiCreateWSDL.sub.heading": "",
+  "Apis.Create.WSDL.Steps.DefaultAPIForm.context": "",
+  "Apis.Create.WSDL.Steps.DefaultAPIForm.name": "",
+  "Apis.Create.WSDL.Steps.DefaultAPIForm.version": "",
+  "Apis.Create.WSDL.Steps.ProvideWSDL.Input.type": "",
+  "Apis.Create.WSDL.Steps.ProvideWSDL.implementation.type": "",
   "Apis.Create.WSDL.Steps.ProvideWSDL.uploaded.file": "",
   "Apis.Create.WSDL.Steps.ProvideWSDL.uploaded.url": "",
   "Apis.Details.APIDefinition.APIDefinition.api.definition": "",
@@ -208,6 +224,8 @@
   "Apis.Details.Endpoints.EndpointOverview.sandbox": "",
   "Apis.Details.Endpoints.Endpoints.cancel": "",
   "Apis.Details.Endpoints.Endpoints.endpoints.header": "",
+  "Apis.Details.Endpoints.Endpoints.managed": "",
+  "Apis.Details.Endpoints.Endpoints.prototyped": "",
   "Apis.Details.Endpoints.Endpoints.save": "",
   "Apis.Details.Endpoints.GeneralConfiguration.Certificates.alias": "",
   "Apis.Details.Endpoints.GeneralConfiguration.Certificates.cancel.button": "",
@@ -242,6 +260,10 @@
   "Apis.Details.Endpoints.LoadBalanceConfig.class.name.for.algorithm": "",
   "Apis.Details.Endpoints.LoadBalanceConfig.session.management": "",
   "Apis.Details.Endpoints.LoadBalanceConfig.session.timeout": "",
+  "Apis.Details.Endpoints.Prototype.InlineEndpoints.script": "",
+  "Apis.Details.Endpoints.Prototype.PrototypeEndpoints.endpoint": "",
+  "Apis.Details.Endpoints.Prototype.PrototypeEndpoints.endpoint.implementation.method": "",
+  "Apis.Details.Endpoints.Prototype.PrototypeEndpoints.inline": "",
   "Apis.Details.EndpointsNew.Endpoints.cancel": "",
   "Apis.Details.EndpointsNew.Endpoints.endpoints.header": "",
   "Apis.Details.EndpointsNew.Endpoints.save": "",
@@ -285,6 +307,9 @@
   "Apis.Details.NewOverview.Configuration.authorization.header.tooltip": "",
   "Apis.Details.NewOverview.Configuration.configuration": "",
   "Apis.Details.NewOverview.Configuration.context": "",
+  "Apis.Details.NewOverview.Configuration.cors.configuration": "",
+  "Apis.Details.NewOverview.Configuration.cors.disabled": "",
+  "Apis.Details.NewOverview.Configuration.cors.enabled": "",
   "Apis.Details.NewOverview.Configuration.created.time": "",
   "Apis.Details.NewOverview.Configuration.default.version": "",
   "Apis.Details.NewOverview.Configuration.description": "",
@@ -412,10 +437,6 @@
   "Apis.Details.Resources.Resources.assign.advanced.throttling.Resource": "",
   "Apis.Details.Resources.Resources.assign.advanced.throttling.perApi": "",
   "Apis.Details.Resources.Resources.assign.advanced.throttling.policies": "",
-  "Apis.Details.Resources.Resources.assign.global.scope.for.api": "",
-  "Apis.Details.Resources.Resources.assign.global.scopes.for.api.button": "",
-  "Apis.Details.Resources.Resources.assign.global.scopes.for.api.input": "",
-  "Apis.Details.Resources.Resources.assign.global.scopes.for.api.title": "",
   "Apis.Details.Resources.Resources.assign.policies": "",
   "Apis.Details.Resources.Resources.cancel": "",
   "Apis.Details.Resources.Resources.delete.selected": "",
