+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/source/src/app/components/Apis/Create/Default/APICreateWrapper.jsx
@@ -21,7 +21,7 @@ import PropTypes from 'prop-types';
 import API from 'AppData/api.js';
 import Alert from 'AppComponents/Shared/Alert';
 import APICreateDefault from './APICreateDefault';
-import APICreateTopMenu from '../Components/APICreateTopMenu';
+
 /**
  * Wrapper for API create component. this wrapper is used when creating
  * an API
