+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/source/src/app/components/Apis/Create/WSDL/ApiCreateWSDL.jsx
@@ -198,7 +198,7 @@ export default function SOAPToREST() {
                                 <Link to='/apis/'>
                                     <Button variant='outlined'>
                                         <FormattedMessage
-                                            id='Apis.Details.Configuration.Configuration.cancel'
+                                            id='Apis.Create.WSDL.ApiCreateWSDL.cancel'
                                             defaultMessage='Cancel'
                                         />
                                     </Button>
