+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/source/src/app/components/Apis/Details/Documents/MarkdownEditor.jsx
@@ -124,10 +124,6 @@ function MarkdownEditor(props) {
     };
 
     const { classes } = props;
-    alert(intl.formatMessage({
-        id: 'documents.markdown.editor.updated.successfully',
-        defaultMessage: 'updated successfully.',
-    }));
     return (
         <div>
             <Button onClick={toggleOpen}>
