+++ features/apimgt/org.wso2.carbon.apimgt.publisher.feature/src/main/resources/publisher-new/source/src/app/data/APIValidation.js
@@ -18,15 +18,10 @@
 import Joi from '@hapi/joi';
 
 const definition = {
-    apiName: Joi.string()
-        .regex(/^[a-zA-Z0-9]{1,30}$/),
-    apiVersion: Joi.string()
-        .regex(/^[a-zA-Z0-9.]{1,30}$/),
-    apiContext: Joi.string()
-        .regex(/^[a-zA-Z0-9/]{1,30}$/),
-    url: Joi.string()
-        .uri(),
+    apiName: Joi.string().regex(/^[a-zA-Z0-9]{1,30}$/),
+    apiVersion: Joi.string().regex(/([\][{}()`~!@#;%^&*+=|\\<>"'/,])/),
+    apiContext: Joi.string().regex(/^[a-zA-Z0-9{}/]{1,30}$/),
+    url: Joi.string().uri(),
 };
-
 
 export default definition;
